back to integration within init
