+++ src/zenml/integrations/gcp/__init__.py
@@ -11,6 +11,26 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
-from zenml.integrations.gcp.integration import GcpIntegration
+from zenml.integrations.integration import Integration
+from zenml.utils.source_utils import import_class_by_path
 
-GcpIntegration.check_installation()
+
+class GcpIntegration(Integration):
+    NAME = "gcp"
+    REQUIREMENTS = ["gcsfs"]
+
+    @classmethod
+    def activate(cls):
+        zen_gcs = import_class_by_path(
+            "zenml.integrations.gcp.io.gcs_plugin.ZenGCS"
+        )
+
+        # TODO: [MEDIUM] we need to implement our own registry
+        from tfx.dsl.io import filesystem_registry
+
+        filesystem_registry.DEFAULT_FILESYSTEM_REGISTRY.register(
+            zen_gcs, priority=15
+        )
+
+        import_class_by_path(
+            "zenml.integrations.gcp.artifact_stores.gcp_artifact_store.GCPArtifactStore")
